deploy/gcp/requirements.txt
# GCP Worker dependencies
flask>=3.0.0
gunicorn>=21.0.0
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.18.0
google-cloud-secret-manager>=2.16.0

